# BigInteger: Replace ==ZERO by isZero.