@glureau 2024-06-12
Thank **you** for all your work! Great library and nice to have such a good support and response time for an open source project. Enjoy your days off!

PS: I've published a 0.3.10-SNAPSHOT2 on my fork https://github.com/users/glureau/packages?repo_name=kotlin-multiplatform-bignum for my own needs with all the improvements if anyone else is interested to test it, but I recommend to wait a bit more if you can ;)